--- lib/common/auth/cubit/auth_state.dart
@@ -1,13 +1,0 @@
-class AuthState {
-  AuthState init() {
-    return AuthState();
-  }
-
-  AuthState clone() {
-    return AuthState();
-  }
-}
-
-class RoleSelected extends AuthState{}
-class GetNumberState extends AuthState{}
-class GetOtpState extends AuthState{}
