+++ lib/common/const/export.dart
@@ -7,3 +7,4 @@ export 'package:ride_options_2/routes/app_route.dart';
 export '../theme/app_theme.dart';
 export 'package:flutter_localizations/flutter_localizations.dart';
 export 'package:flutter_gen/gen_l10n/app_localizations.dart';
+export 'package:flutter_svg/svg.dart';
