+++ lib/common/localization/cubit/localization_cubit.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class LocalizationCubit extends Cubit<Locale> {
-  LocalizationCubit() : super(const Locale('ur')); // Default locale is English
+  LocalizationCubit() : super(const Locale('en')); // Default locale is English
 
   void setEnglishLocale() {
     emit(const Locale('en'));
