--- lib/common/onboarding/cubits/onboard_state.dart
@@ -1,12 +1,0 @@
-class OnboardState {
-  OnboardState init() {
-    return InitState();
-  }
-  OnboardState clone() {
-    return OnboardState();
-  }
-}
-
-class InitState extends OnboardState {}
-
-class ChangeScreen extends OnboardState {}
