+++ lib/common_features/auth/presentation/cubits/auth_state.dart
@@ -1,0 +1,13 @@
+class AuthState {
+  AuthState init() {
+    return AuthState();
+  }
+
+  AuthState clone() {
+    return AuthState();
+  }
+}
+
+class RoleSelected extends AuthState{}
+class GetNumberState extends AuthState{}
+class GetOtpState extends AuthState{}
