+++ lib/common_features/onboarding/cubits/onboard_state.dart
@@ -1,0 +1,12 @@
+class OnboardState {
+  OnboardState init() {
+    return InitState();
+  }
+  OnboardState clone() {
+    return OnboardState();
+  }
+}
+
+class InitState extends OnboardState {}
+
+class ChangeScreen extends OnboardState {}
