+++ lib/driver_features/newFeature/data/new.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class NewDart extends StatelessWidget {
+  const NewDart({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
