+++ lib/l10n/app_en.arb
@@ -1,6 +1,7 @@
 {
   "rideOptions": "Ride Options",
   "pakistan": "Pakistan",
+  "appFontFamily":"Nunito",
   "skip": "Skip",
   "anywhereYouAre": "Anywhere you are",
   "sellHouseEasilly": "Sell houses easily with the help of Listenoryx and to make this line bigI am writing more.",
